perf(daemon): narrow protocol validation to single-pass boundary checks
Cross-process boundaries keep full validation; same-process exits stop
re-validating self-produced data. isJsonObject and rejectSecretKeys judge
only their own layer, since every inbound payload reached us through
JSON.parse. repo.tasks.list exits check the served shape while the client
parse keeps the full row walk; observe.tail drops its second payload
validation behind the RPC entry check; the agent runtime stream hub stops
re-validating its own published events and compares buffered cursors
numerically; action contract regexes compile once via a memo. The
workspace-summary validator moves to its own module to hold the
file-complexity ratchet.
