+++ packages/daemon/src/protocol/daemon-protocol-validate-results.ts
@@ -26,11 +26,8 @@ import {
   validateDaemonDecisionList,
   validateDaemonRelationGraph,
 } from "./daemon-protocol-validate-projections.ts";
-import {
-  queryPageRow,
-  validateDaemonTaskSnapshotList,
-  validateDaemonWorkspaceSummary,
-} from "./daemon-protocol-validate-task.ts";
+import { queryPageRow, validateDaemonTaskSnapshotList } from "./daemon-protocol-validate-task.ts";
+import { validateDaemonWorkspaceSummary } from "./daemon-protocol-validate-workspace.ts";
 import { receiptOutcomeWords } from "./daemon-protocol-vocabulary.ts";
 import { generatedTaskCreateResultFields, generatedWriteReceiptFields } from "./daemon-protocol-commands-task.ts";
 import { isJsonObject, type JsonObject } from "./json-rpc-types.ts";
